Move active_record.rb
