+++ lib/ransack.rb
@@ -19,7 +19,7 @@ end
 require 'ransack/search'
 require 'ransack/ransacker'
 require 'ransack/translate'
-require 'ransack/adapters/active_record'
+require 'ransack/active_record'
 require 'ransack/context'
 
 ActiveSupport.on_load(:action_controller) do
